requirements.txt
Flask==3.0.2
pubnub>=10.4.1
python-dotenv==1.0.1
